+++ content-cache-backends-config/tests/unit/test_state.py
@@ -14,6 +14,8 @@ from errors import ConfigurationError
 from src.state import (
     BACKENDS_CONFIG_NAME,
     BACKENDS_PATH_CONFIG_NAME,
+    HEALTHCHECK_INTERVAL_CONFIG_NAME,
+    HEALTHCHECK_PATH_CONFIG_NAME,
     HOSTNAME_CONFIG_NAME,
     PATH_CONFIG_NAME,
     PROTOCOL_CONFIG_NAME,
@@ -56,6 +58,8 @@ def test_valid_config():
     assert config.protocol == "https"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/"
+    assert config.healthcheck_path == "/healthz"
+    assert config.healthcheck_interval == 2123
     assert config.proxy_cache_valid == ()
 
 
@@ -76,6 +80,9 @@ def test_hostname_with_subdomain():
     assert config.protocol == "https"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/"
+    assert config.healthcheck_path == "/healthz"
+    assert config.healthcheck_interval == 2123
+
     assert config.proxy_cache_valid == ()
 
 
@@ -134,6 +141,7 @@ def test_longer_path():
     charm = MockCharmFactory()
     charm.config[PATH_CONFIG_NAME] = "/path/to/destination/0"
     charm.config[BACKENDS_PATH_CONFIG_NAME] = "/path/to/destination/2"
+    charm.config[HEALTHCHECK_PATH_CONFIG_NAME] = "/path/to/health/2"
 
     config = Configuration.from_charm(charm)
 
@@ -143,6 +151,7 @@ def test_longer_path():
     assert config.protocol == "https"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/path/to/destination/2"
+    assert config.healthcheck_path == "/path/to/health/2"
     assert config.proxy_cache_valid == ()
 
 
@@ -244,6 +253,8 @@ def test_http_protocol():
     assert config.protocol == "http"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/"
+    assert config.healthcheck_path == "/healthz"
+    assert config.healthcheck_interval == 2123
     assert config.proxy_cache_valid == ()
 
 
@@ -375,6 +386,8 @@ def test_valid_proxy_cache_valid(proxy_cache_valid: str):
     assert config.protocol == "https"
     assert config.fail_timeout == "30s"
     assert config.backends_path == "/"
+    assert config.healthcheck_path == "/healthz"
+    assert config.healthcheck_interval == 2123
     assert config.proxy_cache_valid == tuple(json.loads(proxy_cache_valid))
 
 
@@ -396,6 +409,8 @@ def test_configuration_to_data():
         "protocol": "https",
         "fail_timeout": "30s",
         "backends_path": "/",
+        "healthcheck_path": "/healthz",
+        "healthcheck_interval": "2123",
         "proxy_cache_valid": "[]",
     }
 
@@ -432,3 +447,53 @@ def test_configuration_to_json_dumps_error(monkeypatch):
         config.to_integration_data()
 
     assert "Unable to convert configuration to integration data format" in str(err.value)
+
+
+@pytest.mark.parametrize(
+    "bad_value,error_msg",
+    [
+        ("  ", "String should have at least 1 character"),
+        ("{", "Value error, Path contains non-allowed character"),
+    ],
+    ids=["empty", "bad_character"],
+)
+def test_invalid_healthcheck_path(bad_value, error_msg):
+    """
+    arrange: Mock charm with invalid healthcheck path.
+    act: Create the configuration from the charm.
+    assert: Configuration error is raised.
+    """
+    charm = MockCharmFactory()
+    charm.config[HEALTHCHECK_PATH_CONFIG_NAME] = bad_value
+
+    with pytest.raises(ConfigurationError) as err:
+        Configuration.from_charm(charm)
+
+    assert (
+        str(err.value) == f"Config error: ['healthcheck_path = {bad_value.strip()}: {error_msg}']"
+    )
+
+
+@pytest.mark.parametrize(
+    "bad_value,error_msg",
+    [
+        ("  ", "Input should be a valid integer, unable to parse string as an integer"),
+        ("-1", "Input should be greater than 0"),
+        ("0", "Input should be greater than 0"),
+    ],
+    ids=["empty", "negative", "zero"],
+)
+def test_invalid_healthcheck_interval(bad_value, error_msg):
+    """
+    arrange: Mock charm with invalid healthcheck interval.
+    act: Create the configuration from the charm.
+    assert: Configuration error is raised.
+    """
+    charm = MockCharmFactory()
+
+    charm.config[HEALTHCHECK_INTERVAL_CONFIG_NAME] = bad_value
+
+    with pytest.raises(ConfigurationError) as err:
+        Configuration.from_charm(charm)
+
+    assert str(err.value) == f"Config error: ['healthcheck_interval = {bad_value}: {error_msg}']"