+++ content-cache/docs/changelog.md
@@ -6,6 +6,12 @@
 
 - Add grafana dashboard for the displaying metrics from COS integration.
 
+## 2024-12-16
+
+### **Added**
+
+- Support for active healthchecks using LUA scripts. Configured through healthcheck-interval and healthcheck-path.
+
 ## 2024-12-11
 
 ### **Fixed**